Separate the enums from the Sudoku class. Add file for enums and import into Sudoku class
