+++ HelperEnums.py
@@ -1,0 +1,15 @@
+from enum import Enum
+
+
+class Difficulty(Enum):
+    EASY = 51
+    MEDIUM = 53
+    HARD = 56
+
+
+class TermcolorSettings(Enum):
+    RED = 'red'
+    YELLOW = 'yellow'
+    BLUE = 'blue'
+    BOLD = 'bold'
+    DARK = 'dark'
