+++ Sudoku.py
@@ -1,22 +1,8 @@
 import copy
 from pprint import pprint
 import random
-from enum import Enum
 from termcolor import colored
-
-
-class Difficulty(Enum):
-    EASY = 51
-    MEDIUM = 53
-    HARD = 56
-
-
-class TermcolorSettings(Enum):
-    RED = 'red'
-    YELLOW = 'yellow'
-    BLUE = 'blue'
-    BOLD = 'bold'
-    DARK = 'dark'
+from HelperEnums import Difficulty, TermcolorSettings
 
 
 class Sudoku:
